+++ source/extensions/omni.isaac.orbit/omni/isaac/orbit/terrains/__init__.py
@@ -23,8 +23,9 @@ There are two main components in this module:
 """
 
 from .height_field import *  # noqa: F401, F403
-from .terrain_cfg import SubTerrainBaseCfg, TerrainGeneratorCfg, TerrainImporterCfg
 from .terrain_generator import TerrainGenerator
+from .terrain_generator_cfg import SubTerrainBaseCfg, TerrainGeneratorCfg
 from .terrain_importer import TerrainImporter
+from .terrain_importer_cfg import TerrainImporterCfg
 from .trimesh import *  # noqa: F401, F403
 from .utils import color_meshes_by_height, create_prim_from_mesh